Moving definition of EventHorizon
